Start the transition from Jackson to Moshi
The only thing that's supported right now is deserializing Subreddits. I've
completely ditched the Thing class in favor of using its superclass,
RedditObject.

Model classes must be annotated with @RedditModel to be deserialized. For
example, this is the definition for Subreddit:

@RedditModel(kind = KindConstants.SUBREDDIT)
data class Subredit(...)

The deserializer will read the root node and verify its 'kind' attribute with
whatever was specified in @RedditModel and return the deserialized data.